backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-talisman==1.1.0
supabase==2.3.0
openai==1.12.0
assemblyai==0.26.0
deepgram-sdk==3.1.0
pydantic==2.6.0
pydantic-settings==2.1.0
PyJWT==2.8.0
cryptography==42.0.0
python-multipart==0.0.9
tenacity==8.2.3
redis==5.0.1
gunicorn==21.2.0
gevent==24.2.1
gevent-websocket==0.10.1
Werkzeug==3.0.1
requests==2.31.0
python-dotenv==1.0.0
structlog==24.1.0
